source code/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pymysql
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
cloudinary